Add files via upload
Decided to move sorting functions to its own file. There are two .txt files, one for alphabetical sorting and one for high score sorting. There is also a simple scoring system I have implemented in its own file as well.
