+++ RPSFunctions.py
@@ -1,4 +1,6 @@
 import random, sys
+from sort import *
+from score import *
 
 #general function will encapsulate other functions
 #this function will also keep score
@@ -17,10 +19,15 @@ def rps():
             continue
         
         if guess_str == 'q':
+            final = finalscore(wins, losses, ties)
             name = input("Enter your name: ")
-            print_results(wins, losses, ties, name)
-            write('scores.txt', wins, losses, ties, name)
+            print_results(wins, losses, ties, name, final)
+            #scores will be organized in alphabetical order
+            write('scores.txt', wins, losses, ties, name, final)
+            #scores will be organized by score (highest to lowest)
+            write('highscore.txt', wins, losses, ties, name, final)
             alphabetical('scores.txt')
+            highscore('highscore.txt')
             sys.exit()
         
         result, wins, losses, ties = guess(guess_str, wins, losses, ties)
@@ -33,11 +40,12 @@ def rps():
 def inputCheck(input):
     return input in ['0', '1', '2', 'q']
 
-def print_results(wins, losses, ties, name):
+def print_results(wins, losses, ties, name, final):
     print(name)
     print("Wins: " + str(wins))
     print("Losses: " + str(losses))
     print("Ties: " + str(ties))
+    print("Final Score: " + str(final))
 
 def guess(guess_str, wins, losses, ties):
      
@@ -61,17 +69,15 @@ def guess(guess_str, wins, losses, ties):
         losses += 1
 
     return result_message, wins, losses, ties
-
-def score(wins, loses, ties, name):
-    print()
             
-def write(filename, wins, losses, ties, name):
+def write(filename, wins, losses, ties, name, final):
     try: 
         with open(filename, 'a') as f: #'w' for write mode, 'a' for append mode
             f.write(f"Name: {name}\n")
             f.write(f"Wins: {wins}\n")
             f.write(f"Losses: {losses}\n")
             f.write(f"Ties: {ties}\n")
+            f.write(f"Final Score: {final}\n")
             f.write("\n") 
         print(f"Data written to {filename} successfully")
     except Exception as e:
@@ -85,48 +91,4 @@ def load(filename):
     except Exception as e:
         print(f'Error loading data from {file}: {str(e)}')
         return []
-
-def alphabetical(filename):
-    try:
-        with open(filename, 'r') as f:
-            lines = f.readlines()
-
-        # Parse the contents into a list of dictionaries 
-        records = []
-        current_record = {}
-
-        for line in lines:
-            line = line.strip()
-            if line.startswith("Name:"):
-                if current_record: 
-                    records.append(current_record)
-                current_record = {"Name": line.replace("Name: ", "")}
-            elif line.startswith("Wins:"):
-                current_record["Wins"] = int(line.replace("Wins: ",""))
-            elif line.startswith("Losses"):
-                current_record["Losses"] = int(line.replace("Losses: ", ""))
-            elif line.startswith("Ties:"):
-                current_record["Ties"] = int(line.replace("Ties: ", ""))
-
-        if current_record: #Add the last record
-            records.append(current_record)
-
-        #sort the records alphabeticallty by name
-        records.sort(key=lambda x: x["Name"])
-
-        #write the sorted records back to the file
-        with open(filename, 'w') as f:
-            for record in records:
-                f.write(f"Name: {record['Name']}\n")
-                f.write(f"Wins: {record['Wins']}\n")
-                f.write(f"Losses: {record['Losses']}\n")
-                f.write(f"Ties: {record['Ties']}\n")
-                f.write("\n")
-
-        print(f"{filename} reorganized alphabetically by name successfully.")
-
-    except Exception as e:
-        print(f"Error reorganizing {filename}: {str(e)}")
-
-
