+++ main.py
@@ -1,10 +1,6 @@
-#from FirstScriptFunction import *
-#from RockPaperScissors import *
 from RPSFunctions import *
 
 def main():
-    #first_script()
-    #rockpaperscissors()
     rps()
 
 if __name__ == "__main__":
